Remove use of namerefs, so it works with macOS Bash
The Bash on macOS is old, really old, as is well known. It is stuck at 3.x
and therefore does not have the nameref feature. To make the script run on
macOS, it cannot use newer Bash features. So this commit changes the
modern use of a reference to the old way of using 'eval' to change the
value of a variable, the name of which was passed in to the function.